+++ src/operations/manage_sell_offer.js
@@ -1,4 +1,4 @@
-import { Hyper } from 'js-xdr';
+import { Hyper } from '@stellar/js-xdr';
 import xdr from '../xdr';
 /**
  * Returns a XDR ManageSellOfferOp. A "manage sell offer" operation creates, updates, or