+++ test/unit/sorobandata_builder_test.js
@@ -6,17 +6,17 @@ describe('SorobanTransactionData can be built', function () {
   const c = new StellarBase.Contract(contractId);
 
   const sentinel = new xdr.SorobanTransactionData({
+    ext: new xdr.ExtensionPoint(0),
     resources: new xdr.SorobanResources({
       footprint: new xdr.LedgerFootprint({ readOnly: [], readWrite: [] }),
       instructions: 1,
       readBytes: 2,
       writeBytes: 3
     }),
-    ext: new xdr.ExtensionPoint(0),
-    refundableFee: new xdr.Int64(5)
+    resourceFee: new xdr.Int64(5)
   });
 
-  const key = c.getFootprint()[0];
+  const key = c.getFootprint(); // arbitrary key for testing
 
   it('constructs from xdr, base64, and nothing', function () {
     new dataBuilder();
@@ -35,7 +35,7 @@ describe('SorobanTransactionData can be built', function () {
 
   it('sets properties as expected', function () {
     expect(
-      new dataBuilder().setResources(1, 2, 3).setRefundableFee(5).build()
+      new dataBuilder().setResources(1, 2, 3).setResourceFee(5).build()
     ).to.eql(sentinel);
 
     // this isn't a valid param but we're just checking that setters work
@@ -73,8 +73,8 @@ describe('SorobanTransactionData can be built', function () {
   it('makes copies on build()', function () {
     const builder = new dataBuilder();
     const first = builder.build();
-    const second = builder.setRefundableFee(100).build();
+    const second = builder.setResourceFee(100).build();
 
-    expect(first.refundableFee()).to.not.eql(second.refundableFee());
+    expect(first.resourceFee()).to.not.eql(second.resourceFee());
   });
 });